+++ CHANGELOG.md
@@ -7,17 +7,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+
 ### Added
 
 - Added support for A-F to Digits widget https://github.com/Textualize/textual/pull/5094
 - Added `Region.constrain` https://github.com/Textualize/textual/pull/5097
+- Added support for A-F to Digits widget https://github.com/Textualize/textual/pull/5094
 
 ### Changed
 
+- `Screen.ALLOW_IN_MAXIMIZED_VIEW` will now default to `App.ALLOW_IN_MAXIMIZED_VIEW` https://github.com/Textualize/textual/pull/5088
+- Widgets matching `.-textual-system` will now be included in the maximize view by default https://github.com/Textualize/textual/pull/5088
 - Digits are now thin by default, style with text-style: bold to get bold digits https://github.com/Textualize/textual/pull/5094
 - Made `Widget.absolute_offset` public https://github.com/Textualize/textual/pull/5097
 - Tooltips are now displayed directly below the mouse cursor https://github.com/Textualize/textual/pull/5097
 - `Region.inflect` will now assume that margins overlap https://github.com/Textualize/textual/pull/5097
+- `Pilot.click` and friends will now accept a widget, in addition to a selector https://github.com/Textualize/textual/pull/5095
 
 ## [0.82.0] - 2024-10-03